+++ src/visual-feature/vpFeatureMoment.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- * $Id: vpFeatureMoment.cpp 4056 2013-01-05 13:04:42Z fspindle $
+ * $Id: vpFeatureMoment.cpp 4303 2013-07-04 14:14:00Z fspindle $
  *
  * This file is part of the ViSP software.
  * Copyright (C) 2005 - 2013 by INRIA. All rights reserved.
@@ -64,7 +64,7 @@ void vpFeatureMoment::init (){
 	 * This partly explains why vpFeatureMomentBasic cannot be used directly as a visual feature.
 	 */
     if(this->moment!=NULL)
-        dim_s = this->moment->get().size();
+        dim_s = (unsigned int)this->moment->get().size();
     else
         dim_s = 0;
 
@@ -159,7 +159,7 @@ void vpFeatureMoment::update (double A, double B, double C){
     }
     nbParameters = 1;
     if(this->moment!=NULL){
-        dim_s = this->moment->get().size();
+        dim_s = (unsigned int)this->moment->get().size();
 
         s.resize(dim_s);
